pyproject.toml
[project]
name = "looprock-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.10",
    "aiosqlite>=0.20.0",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.12.3",
    "feedparser>=6.0.11",
    "niquests>=3.11.2",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2024.2",
    "tomli>=2.2.1",
]
